Make CUDA and Term weak dependencies on Julia >= 1.9
