+++ Project.toml
@@ -1,6 +1,6 @@
 name = "XGBoost"
 uuid = "009559a3-9522-5dbb-924b-0b6ed2b22bb9"
-version = "2.2.5"
+version = "2.3.0"
 
 [deps]
 AbstractTrees = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"
@@ -14,8 +14,15 @@ SparseMatricesCSR = "a0a7dd2c-ebf4-11e9-1f05-cf50bc540ca1"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 Term = "22787eb5-b846-44ae-b979-8e399b8463ab"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 XGBoost_jll = "a5c6f535-4255-5ca2-a466-0e519f119c46"
+
+[weakdeps]
+CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
+Term = "22787eb5-b846-44ae-b979-8e399b8463ab"
+
+[extensions]
+XGBoostCUDAExt = "CUDA"
+XGBoostTermExt = "Term"
 
 [compat]
 AbstractTrees = "0.4"
@@ -30,8 +37,10 @@ XGBoost_jll = "1.7.2"
 julia = "1.6"
 
 [extras]
+CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+Term = "22787eb5-b846-44ae-b979-8e399b8463ab"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Random", "Test"]
+test = ["CUDA", "Random", "Term", "Test"]
