+++ src/XGBoost.jl
@@ -9,8 +9,6 @@ using AbstractTrees
 using OrderedCollections
 using JSON3
 using Tables
-using Term
-using CUDA
 using Statistics: mean, std
 
 using Base: @propagate_inbounds
@@ -49,5 +47,9 @@ include("introspection.jl")
 include("show.jl")
 include("defaultparams.jl")
 
+if !isdefined(Base, :get_extensions)
+    include("../ext/XGBoostCUDAExt.jl")
+    include("../ext/XGBoostTermExt.jl")
+end
 
 end # module XGBoost
